Move rehydrate function
