+++ src/lite/index.js
@@ -119,6 +119,27 @@ export const addPx = (prop, value) => {
   return value + 'px'
 }
 
+export const rehydrate = (css) => {
+  let dec
+
+  while (dec = RULE_REG.exec(css)) {
+    const media = dec[2] || ''
+    const className = dec[3]
+    const pseudo = dec[4] || ''
+    const key = camel(dec[5])
+    const val = removePx(dec[6])
+    const id = key + val + media + pseudo
+
+    cache[id] = className
+  }
+}
+
+const RULE_REG = /((@media[^{]+){)?.([^:{]+)(:[^{]+)?{([^:]+):([^}]+)}}?/g
+
+const camel = (str) => str.replace(/-[a-z]/g, (g) => g[1].toUpperCase())
+
+const removePx = (str) => str.replace(/px$/, '')
+
 const unitlessProps = [
   'animationIterationCount',
   'boxFlex',
@@ -152,6 +173,8 @@ const unitlessProps = [
 
 cxs.css = css
 cxs.reset = reset
+cxs.cache = cache
+cxs.rehydrate = rehydrate
 
 export default cxs
